Redo goals table migrations
I really messed these up when I initially created this, and it was
prohibitively hard to correct those mistakes in a way that was
reversible.

Note: If you had a database before this, you'll need to completely
delete your docker volume and recreate it. You might need to delete
node_modules and reinstall yarn packages.